src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/import-rule-in-shadow-tree-needs-document-style-recalc-expected.txt

CONSOLE ERROR: line 6: Uncaught TypeError: host.createShadowRoot is not a function
